pyproject.toml
[project]
name          = "ruamel_yaml_line_info"
version       = "0.1.0"
description   = "Line information for ruamel.yaml."
readme        = "README.md"
#authors       = ["Nils Homer <nilshomer@gmail.com>"]
license       = "MIT"
classifiers   = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = [ "dependencies", "requires-python" ]

[project.urls]
homepage      = "https://github.com/nh13/ruamel_yaml_line_info"
repository    = "https://github.com/nh13/ruamel_yaml_line_info"
documentation = "https://github.com/nh13/ruamel_yaml_line_info"
"Bug Tracker" = "https://github.com/nh13/ruamel_yaml_line_info/issues"

[tool.poetry]
include       = ["LICENSE"]

[tool.poetry.dependencies]
python    = "^3.11"
ruamel-yaml = "^0.18.10"

[tool.poetry.group.dev.dependencies]
mypy        = "^1.13.0"
pytest      = "^8.3.3"
pytest-cov  = "^5.0.0"
ruff        = "0.7.4"

[tool.poetry.group.ipython]
optional = true

[tool.poetry.group.ipython.dependencies]
ipython = "^8.24.0"
ipdb    = "^0.13.13"

[tool.poe]
poetry_command = "task"

[tool.poe.tasks]
fix-format = "ruff format"
fix-lint   = "ruff check --fix"

fix-all.ignore_fail = "return_non_zero"
fix-all.sequence    = [
    "fix-format",
    "fix-lint"
]

check-lock   = "poetry check --lock"
check-format = "ruff format --check --diff"
check-lint   = "ruff check"
check-tests  = "pytest"
check-typing = "mypy"

check-all.ignore_fail = "return_non_zero"
check-all.sequence    = [
    "check-lock",
    "check-format",
    "check-lint",
    "check-tests",
    "check-typing"
]

fix-and-check-all.ignore_fail = "return_non_zero"
fix-and-check-all.sequence    = [
    "fix-all",
    "check-all"
]

[build-system]
requires      = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files                       = ["./"]
strict_optional             = true
strict_equality             = true
check_untyped_defs          = true
disallow_incomplete_defs    = true
disallow_untyped_calls      = true
disallow_untyped_decorators = true
disallow_untyped_defs       = true
no_implicit_optional        = true
warn_no_return              = true
warn_redundant_casts        = true
warn_return_any             = true
warn_unreachable            = true
warn_unused_configs         = true
warn_unused_ignores         = true
enable_error_code           = [
    "ignore-without-code",
    "possibly-undefined",
]
exclude                     = [
    "docs/",
    "site/",
]

[[tool.mypy.overrides]]
module = "ruamel"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts    = [
    "--color=yes",
    "--import-mode=importlib",
	"--cov=ruamel_yaml_line_info",
	"--cov-report=xml",
	"--cov-branch",
]

[tool.ruff]
line-length    = 100
target-version = "py311"
output-format  = "full"
preview        = true

[tool.ruff.lint]
select    = [
    "ARG",    # Unused arguments
    "C901",   # McCabe complexity
    "B",      # bugbear
    "D",      # pydocstyle (docstrings. We have the "google" convention enabled)
    "D204",   # Blank line between class docstring and first (__init__) method
    "D213",   # Summary line should be located on the line after opening quotes
    "E",      # pycodestyle errors
    "LOG",    # flake8-logging
    "LOG015", # (preview rule) Prohibit calls to the root logger
    "F",      # pyflakes
    "I",      # isort
    "N",      # PEP8 naming
    "W",      # pycodestyle warnings
    "Q",      # flake8-quotes
]
ignore    = [
    "E203",
    "E701",
    "D212",  # summary line should be located on the same line as opening quotes
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
]
unfixable = ["B"]
# NB: only preview rules explicitly selected above (e.g. LOG015) will be enforced
preview = true
explicit-preview-rules = true

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.pydocstyle]
convention = "google"
